add TestNode.wrap method
This is a bit more explicit. Sometimes we have a String, others we start
with a `Capybara::Node::Simple`, and yet others we're dealing with a
`Nokogiri` element. For the cases when we're already starting with a
`Simple` we can call `new`, but for cases where we need to wrap it we
can call `wrap`.